clean-up

+++ Answer/Inc/ControlSensor.h
@@ -6,7 +6,7 @@
 
 typedef struct
 {
-  float steeringAngle;
+  double steeringAngle;
 } ControlSensor;
 
 #endif
